requirements.txt
numpy
pandas
python-dateutil
yfinance
beautifulsoup4
lxml
nltk
vaderSentiment
wordcloud
requests
retrying
tqdm
newsapi-python
configparser
python-dotenv
virtualenv
matplotlib
plotly
seaborn
streamlit
diskcache
openpyxl
pyarrow
sqlalchemy
jsonschema
scikit-learn
dvc
cryptography
streamz
ipykernel
notebook
black
flake8
pytest
gitpython
dash
Pillow
weasyprint
jinja2
kaleido
